+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * `delphin.itsdb.ItsdbProfile.exists()` (#112)
 * `delphin.itsdb.ItsdbProfile.size()` (#112)
 * `--in-place` option to the `delphin mkprof` command (#109)
+* `delphin.derivation.UdfNode.preterminals()` (#105)
+* `delphin.derivation.UdfNode.terminals()` (#105)
 
 ### Changed